feat: add analytics new func
